added i10_id and test
