+++ tests/devices/i10/lookupTables/IDEnergy2GapCalibrations.csv
@@ -1,0 +1,53 @@
+Source,Mode,MinEnergy,MaxEnergy,b,1st-order,2nd-order,3rd-order,4th-order,5th-order,6th-order,7th-order
+idu,lh,255.3,513.28,3.88755,0.0641791,-7.52562e-05,4.33435e-08,0,0,0,0
+idu,lh,513.28,771.545,8.1496,0.0393435,-2.61385e-05,1.04394e-08,0,0,0,0
+idu,lh,771.545,1027.52,7.9205,0.0393892,-2.51296e-05,9.55013e-09,0,0,0,0
+idu,lh,1027.52,1282.59,6.60845,0.0430451,-2.84144e-05,1.04932e-08,0,0,0,0
+idu,lh,1282.59,1700.0,-160.378,0.421009,-0.000314196,8.26909e-08,0,0,0,0
+idu,lv,521.99,734.983,6.96008,0.0202971,-6.57463e-06,1.67544e-09,0,0,0,0
+idu,lv,734.983,932.69,7.8749,0.0181492,-5.82037e-06,2.32314e-09,0,0,0,0
+idu,lv,932.69,1129.16,3.24578,0.0331944,-2.22386e-05,8.33737e-09,0,0,0,0
+idu,lv,1129.16,1325.74,-14.1739,0.0777255,-6.01726e-05,1.91056e-08,0,0,0,0
+idu,lv,1325.74,1700.0,-172.027,0.423051,-0.000312082,8.03864e-08,0,0,0,0
+idu,pc,418.566,637.695,5.13046,0.0368795,-2.99885e-05,1.35966e-08,0,0,0,0
+idu,pc,637.695,855.483,6.88328,0.0285808,-1.67934e-05,6.54757e-09,0,0,0,0
+idu,pc,855.483,1071.32,6.03968,0.0309807,-1.90157e-05,7.2133e-09,0,0,0,0
+idu,pc,1071.32,1287.03,-5.94187,0.0630626,-4.76693e-05,1.57511e-08,0,0,0,0
+idu,pc,1287.03,1700.0,-97.3851,0.270486,-0.00020465,5.53938e-08,0,0,0,0
+idu,nc,426.626,646.014,5.26037,0.0361496,-2.86867e-05,1.28076e-08,0,0,0,0
+idu,nc,646.014,863.674,6.39722,0.030442,-1.92019e-05,7.57955e-09,0,0,0,0
+idu,nc,863.674,1079.84,7.21445,0.0273263,-1.52587e-05,5.92247e-09,0,0,0,0
+idu,nc,1079.84,1295.74,-7.52454,0.0668457,-5.06815e-05,1.65398e-08,0,0,0,0
+idu,nc,1295.74,1700.0,-104.872,0.28583,-0.000215111,5.77565e-08,0,0,0,0
+idu,lh3,769.119,1028.01,1.83429,0.0274816,-1.4457e-05,3.62736e-09,0,0,0,0
+idu,lh3,1028.01,1286.16,3.87065,0.0214904,-8.56945e-06,1.69557e-09,0,0,0,0
+idu,lh3,1286.16,1544.17,6.38981,0.0157398,-4.20572e-06,5.94806e-10,0,0,0,0
+idu,lh3,1544.17,1794.62,-11.6837,0.0499948,-2.58804e-05,5.17381e-09,0,0,0,0
+idu,lh3,1794.62,2100,-156.01,0.268796,-0.000135379,2.32221e-08,0,0,0,0
+idu,la,654,1700.0,-43.829175,0.368501,-0.0010666011,1.7879222e-06,-1.7983774e-09,1.0820994e-12,-3.6007756e-16,5.1239781e-20
+idd,lh,249.772,487.644,4.12064,0.0680339,-8.56005e-05,5.23532e-08,0,0,0,0
+idd,lh,487.644,731.037,6.81039,0.0492164,-4.20426e-05,1.89432e-08,0,0,0,0
+idd,lh,731.037,980.599,12.0757,0.027081,-1.05292e-05,3.77061e-09,0,0,0,0
+idd,lh,980.599,1236.31,5.79522,0.0473828,-3.24713e-05,1.16963e-08,0,0,0,0
+idd,lh,1236.31,1700.0,-61.7239,0.206534,-0.000157637,4.4545e-08,0,0,0,0
+idd,lv,496.652,689.325,3.53496,0.0400226,-3.86443e-05,1.90295e-08,0,0,0,0
+idd,lv,689.325,886.548,6.65066,0.0247265,-1.35913e-05,5.35915e-09,0,0,0,0
+idd,lv,886.548,1088.57,1.29591,0.0412981,-3.06327e-05,1.1181e-08,0,0,0,0
+idd,lv,1088.57,1294.24,-8.58906,0.0655113,-5.01767e-05,1.63633e-08,0,0,0,0
+idd,lv,1294.24,1700.0,-91.2008,0.250993,-0.00018901,5.10085e-08,0,0,0,0
+idd,pc,385.505,596.042,4.67408,0.0433433,-4.29176e-05,2.21358e-08,0,0,0,0
+idd,pc,596.042,810.876,6.97271,0.0307363,-1.97097e-05,7.83367e-09,0,0,0,0
+idd,pc,810.876,1030.87,8.69789,0.0244055,-1.19019e-05,4.5975e-09,0,0,0,0
+idd,pc,1030.87,1255.31,6.55313,0.0324129,-2.13844e-05,8.21876e-09,0,0,0,0
+idd,pc,1255.31,1700.0,-49.1463,0.164373,-0.000125734,3.57615e-08,0,0,0,0
+idd,nc,394.958,605.487,4.82301,0.0422041,-4.02467e-05,2.01446e-08,0,0,0,0
+idd,nc,605.487,820.328,7.1421,0.0300373,-1.87797e-05,7.4329e-09,0,0,0,0
+idd,nc,820.328,1040.05,8.04047,0.0266286,-1.44209e-05,5.55491e-09,0,0,0,0
+idd,nc,1040.05,1264.12,-2.79321,0.0572147,-4.32987e-05,1.46731e-08,0,0,0,0
+idd,nc,1264.12,1700.0,-75.8958,0.22423,-0.00017034,4.68429e-08,0,0,0,0
+idd,lh3,710.953,951.552,1.71019,0.0304018,-1.78548e-05,4.92026e-09,0,0,0,0
+idd,lh3,951.552,1193.71,4.05097,0.0230031,-1.00382e-05,2.15916e-09,0,0,0,0
+idd,lh3,1193.71,1438.26,7.78066,0.013952,-2.70169e-06,1.72396e-10,0,0,0,0
+idd,lh3,1438.26,1684.34,6.84516,0.0161913,-4.51536e-06,6.65626e-10,0,0,0,0
+idd,lh3,1684.34,2100,49.5656,-0.0582239,3.85587e-05,-7.61825e-09,0,0,0,0
+idd,la,610,1200,800.865724,-7.60298294,0.0309135438,-6.88548924e-05,9.10074539e-08,-7.14462774e-11,3.08703041e-14,-5.66653297e-18
