+++ tests/devices/i10/lookupTables/IDEnergy2PhaseCalibrations.csv
@@ -1,0 +1,29 @@
+Source,Mode,MinEnergy,MaxEnergy,b,1st-order,2nd-order,3rd-order,4th-order,5th-order,6th-order,7th-order
+idu,pc,418.566,637.695,13.4873,0.00640997,-6.03096E-06,2.77583E-09,0,0,0,0
+idu,pc,637.695,855.483,14.4496,0.00216711,2.76266E-07,-3.9213E-10,0,0,0,0
+idu,pc,855.483,1071.32,14.27,0.00321593,-1.46775E-06,5.00523E-10,0,0,0,0
+idu,pc,1071.32,1287.03,15.739,-0.000289772,1.28573E-06,-2.10213E-10,0,0,0,0
+idu,pc,1287.03,1700,11.4359,0.0105901,-7.83598E-06,2.32768E-09,0,0,0,0
+idu,nc,426.626,646.014,-13.8004,-0.00463626,2.71849E-06,-7.28072E-10,0,0,0,0
+idu,nc,646.014,863.674,-14.0729,-0.00364046,1.64441E-06,-4.38739E-10,0,0,0,0
+idu,nc,863.674,1079.84,-13.0961,-0.00679536,5.09164E-06,-1.71735E-09,0,0,0,0
+idu,nc,1079.84,1295.74,-11.888,-0.0095199,7.02495E-06,-2.13009E-09,0,0,0,0
+idu,nc,1295.74,1700,-7.59802,-0.018822,1.37127E-05,-3.72318E-09,0,0,0,0
+idu,la,654,1700,3.1537295,-0.11906434,0.00034602416,-5.6632637E-07,5.5036409E-10,-3.1630855E-13,9.9437208E-17,-1.3207099E-20
+idd,pc,385.505,596.042,13.2841,0.00827853,-9.91799E-06,5.3993E-09,0,0,0,0
+idd,pc,596.042,810.876,14.0914,0.00400033,-2.22442E-06,7.20919E-10,0,0,0,0
+idd,pc,810.876,1030.87,14.8875,0.00143318,5.1847E-07,-2.49769E-10,0,0,0,0
+idd,pc,1030.87,1255.31,15.4992,0.000445666,7.26373E-07,-8.06145E-11,0,0,0,0
+idd,pc,1255.31,1700,14.46,0.00382958,-2.6632E-06,9.97774E-10,0,0,0,0
+idd,nc,394.958,605.487,-13.6686,-0.00588126,4.99786E-06,-2.07139E-09,0,0,0,0
+idd,nc,605.487,820.328,-14.0886,-0.00394451,2.05578E-06,-6.01609E-10,0,0,0,0
+idd,nc,820.328,1040.05,-13.3465,-0.00643408,4.87638E-06,-1.68506E-09,0,0,0,0
+idd,nc,1040.05,1264.12,-12.8173,-0.0075141,5.46726E-06,-1.72449E-09,0,0,0,0
+idd,nc,1264.12,1700,-9.07942,-0.0157672,1.14767E-05,-3.16431E-09,0,0,0,0
+idd,la,610,1200,1400.13845,-13.0099589,0.0518055008,-0.000113471444,1.47749738E-07,-1.14420814E-10,4.88195989E-14,-8.85680201E-18
+idd,lh,249.772,1700,0,0,0,0,0,0,0,0
+idu,lh,255.3,1700,0,0,0,0,0,0,0,0
+idd,lv,496.652,1700,24,0,0,0,0,0,0,0
+idu,lv,521.99,1700,24,0,0,0,0,0,0,0
+idd,lh3,710.953,2100,0,0,0,0,0,0,0,0
+idu,lh3,769.119,2100,0,0,0,0,0,0,0,0
